Rearrange exclude code to filter source and target URLs
Move exclude code handler to mirror-url and mirror-main

Fixes #2319
